implementation of attacks complete

+++ src/attacks/controllers/floodingAttack/NA_FloodingAttack.cc
@@ -12,8 +12,7 @@ cMessage *NA_FloodingAttack::generateAttackMessage(const char* cmsg) {
     // Specific message for the specifics hacked modules.
     NA_FloodingMessage *msg = new NA_FloodingMessage(cmsg);
     msg->setFloodingGradeIndicator(
-            par("floodingGradeIndicator").doubleValue());
-
+            par("floodingGradeIndicator"));
     return msg;
 }
